Refactor components for improved responsiveness, and enhance code formatting

+++ backend/src/utils/cloudinary.js
@@ -1,29 +1,33 @@
-import {v2 as cloudinary} from 'cloudinary'
+import { v2 as cloudinary } from 'cloudinary'
 import fs from "fs";
 import dotenv from "dotenv";
 
-dotenv.config(); 
+dotenv.config();
 cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET
 })
 
-const uploadOnCloudinary=async function(localFilePath){
+const uploadOnCloudinary = async function (localFilePath) {
   try {
-      if(!localFilePath) return null;
-      const response=await cloudinary.uploader.upload(localFilePath,{
-          resource_type:"auto"
-      })
-       console.log("Cloudinary upload success:", response.secure_url);
-      fs.unlinkSync(localFilePath)
-      return response;
+    if (!localFilePath) return null;
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto"
+    })
+    console.log("Cloudinary upload success:", response.secure_url);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+    return response;
 
   } catch (error) {
-    console.error("Cloudinary Upload Error:", error); 
-    fs.unlinkSync(localFilePath)
-    return null;    
+    console.error("Cloudinary Upload Error:", error);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+    return null;
   }
 }
 
-export {uploadOnCloudinary}
+export { uploadOnCloudinary }
